prevent author from being deleted if has books

+++ web/src/main/kotlin/com/demo/project/librarystore/service/AuthorService.kt
@@ -1,10 +1,12 @@
 package com.demo.project.librarystore.service
 
 import com.demo.project.librarystore.exception.IdAlreadyExistsException
+import com.demo.project.librarystore.exception.NotSupportedOperation
 import com.demo.project.librarystore.exception.ResourceNotFoundException
 import com.demo.project.librarystore.model.AuthorModel
 import com.demo.project.librarystore.model.toModel
 import com.demo.project.librarystore.repository.AuthorRepository
+import com.demo.project.librarystore.repository.BookRepository
 import org.springframework.stereotype.Service
 import java.lang.RuntimeException
 import java.time.LocalDate
@@ -13,6 +15,7 @@ import org.springframework.dao.DuplicateKeyException
 @Service
 class AuthorService(
     private val authorRepository: AuthorRepository,
+    private val bookRepository: BookRepository,
 ) {
     fun getAllAuthors(): List<AuthorModel> {
         return authorRepository.getAllAuthors().map { it.toModel() }
@@ -50,6 +53,9 @@ class AuthorService(
     }
 
     fun deleteAuthorById(id: Int) {
+        if (bookRepository.isThereAssociatedBook(id)) {
+            throw NotSupportedOperation("Author has associated books.")
+        }
         val deletedCount = authorRepository.deleteAuthor(id)
         if (deletedCount == 0) {
             throw ResourceNotFoundException("Author not found.")
